requirements.txt
pandas==2.3.2
python-dotenv==1.1.1
requests==2.32.5
xlsxwriter==3.2.5